Moved python to be loaded separately
