+++ src/api.js
@@ -1,27 +1,17 @@
+import pythonAPI from './python/api.py'
 
-export default function handleUploaded903Data(uploadedFiles) {
+export async function handleUploaded903Data(uploadedFiles) {
     const pyodide = window.pyodide;
 
     console.log('Passing uploaded data to Pyodide...')
     pyodide.globals.set("uploaded_files", uploadedFiles)
-    pyodide.runPython(`
-    import pandas as pd
-    from io import StringIO
 
-    def get_file_type(df):
-        if 'UPN' in df.columns:
-            return 'Header'
-        elif 'DECOM' in df.columns:
-            return 'Episodes'
+    const script = await (await fetch(pythonAPI)).text();
+    pyodide.runPython(script)
 
-    files = {}
+    return pyodide.globals.get("js_files").toJs()
+}
 
-    for file_name, file in uploaded_files.to_py().items():
-        csv_file = StringIO(file)
-        df = pd.read_csv(csv_file)
+export function get903Errors(df) {
 
-        files[get_file_type(df)] = [t._asdict() for t in df.itertuples(index=False)]
-    `)
-
-    return pyodide.globals.get("files").toJs()
 }
